Add connection test for Substrate
